[Feature #156113633] User recipe edit and delete features.
EDIT

Given a user is on the dashboard and is viewing their recipes
When they click Edit Recipe on any their categories
Then they should be receive a prompt with the values of the recipe
When the modify the values and click Update Recipe
Then their recipe should be updated successfully.

DELETE

Given a user is on the dashboard and is viewing their recipes
When they click Delete Recipe on any their categories
Then they should be receive a prompt to confirm that they indeed want to delete the selected recipe
When the click Confirm Delete
Then their recipe should be deleted successfully.

OTHER

This are other minor changes made to the add and view for categories and recipes following API response
refactors.